requirements.txt
requests
scikit-learn==1.6.1
pandas==2.3.2
google-cloud-vision==3.10.2
